+++ Usage/src/main/java/io/pivotal/tola/cfapi/usage/model/OrgUsage.java
@@ -65,7 +65,7 @@ public class OrgUsage {
     public String getTotalGbPerAis(){
 
         DecimalFormat df = new DecimalFormat("#.###");
-        return df.format(this.getTotalMbPerAis() / 1024);
+        return df.format( (double) this.getTotalMbPerAis() / 1024);
 
     }
