+++ Usage/src/main/java/io/pivotal/tola/cfapi/usage/model/SpaceUsage.java
@@ -28,7 +28,7 @@ public class SpaceUsage {
     public String getTotalGbPerAis(){
 
         DecimalFormat df = new DecimalFormat("#.###");
-        return df.format(this.getTotalMbPerAis() / 1024);
+        return df.format( (double)this.getTotalMbPerAis() / 1024);
 
     }
